IGNITE-20226 Sql. Fix muted tests after disallowed char -> varbinary casting
