+++ modules/runner/src/integrationTest/java/org/apache/ignite/internal/sql/engine/datatypes/uuid/ItUuidExpressionTest.java
@@ -17,6 +17,7 @@
 
 package org.apache.ignite.internal.sql.engine.datatypes.uuid;
 
+import static org.apache.ignite.lang.IgniteStringFormatter.format;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -25,10 +26,14 @@ import java.util.UUID;
 import org.apache.ignite.internal.sql.engine.datatypes.DataTypeTestSpecs;
 import org.apache.ignite.internal.sql.engine.datatypes.tests.BaseExpressionDataTypeTest;
 import org.apache.ignite.internal.sql.engine.datatypes.tests.DataTypeTestSpec;
+import org.apache.ignite.internal.sql.engine.datatypes.tests.TestTypeArguments;
 import org.apache.ignite.internal.sql.engine.type.UuidType;
 import org.apache.ignite.lang.IgniteException;
+import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 
 /**
  * Tests for expressions for {@link UuidType UUID data type}.
@@ -63,6 +68,31 @@ public class ItUuidExpressionTest extends BaseExpressionDataTypeTest<UUID> {
     public void testInvalidUuidStringInDynamicParams() {
         IgniteException t = assertThrows(IgniteException.class, () -> runSql("SELECT ?::UUID", "00000"));
         assertThat(t.getMessage(), containsString("Invalid UUID string"));
+    }
+
+    /**
+     * {@code COALESCE} operator does not allow different types in its arguments.
+     */
+    @ParameterizedTest
+    @MethodSource("convertedFrom")
+    public void testCoalesceMissingTypesIsIllegal(TestTypeArguments arguments) {
+        IgniteException t = assertThrows(IgniteException.class, () -> {
+            checkQuery(format("SELECT COALESCE($0, {})", arguments.valueExpr(0))).check();
+        });
+
+        assertThat(t.getMessage(), Matchers.containsString("Illegal mixing of types in CASE or COALESCE statement"));
+    }
+
+    /** Data type from string. **/
+    @Test
+    public void testCastFromString() {
+        UUID value = dataSamples.values().get(0);
+        String stringValue = testTypeSpec.toStringValue(value);
+
+        checkQuery(format("SELECT CAST('{}' AS <type>)", stringValue)).returns(value).check();
+
+        // PG style cast
+        checkQuery(format("SELECT '{}'::<type>", stringValue)).returns(value).check();
     }
 
     /** {@inheritDoc} **/
